# docs/architecture/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/api
  description: Generated server url
...
paths:
...
  /appointments:
    get:
      tags:
      - Appointments
      summary: List all appointments
      description: Retrieves a paginated list of appointments with optional search
      operationId: listAppointments
      parameters:
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: perPage
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
        example: 10
      - name: search
        in: query
        description: Search terms
        required: false
        schema:
          type: string
          default: ""
        example: CRECHE
      - name: sort
        in: query
        description: Sort field
        required: false
        schema:
          type: string
          default: startAt
        example: startAt
      - name: direction
        in: query
        description: Sort direction (asc/desc)
        required: false
        schema:
          type: string
          default: asc
        example: asc
      responses:
        "200":
          description: Appointments retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationAppointmentResponse"
    post:
      tags:
      - Appointments
      summary: Create a new appointment
      description: Creates a new appointment for a pet
      operationId: createAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
        required: true
      responses:
        "422":
          description: Validation error
        "201":
          description: Appointment created successfully
          headers:
            Location:
              description: URI of the created appointment
              style: simple
        "400":
          description: Invalid input
  /appointments/{id}/status:
    patch:
      tags:
      - Appointments
      summary: Change appointment status
      description: Changes the lifecycle status of an appointment
      operationId: changeAppointmentStatus
      parameters:
      - name: id
        in: path
        description: Appointment ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeAppointmentStatusRequest"
        required: true
      responses:
        "422":
          description: Invalid status transition
        "200":
          description: Status changed successfully
        "404":
          description: Appointment not found
  /appointments/{id}:
    get:
      tags:
      - Appointments
      summary: Get appointment by ID
      description: Retrieves an appointment by its ID
      operationId: getAppointmentById
      parameters:
      - name: id
        in: path
        description: Appointment ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
    delete:
      tags:
      - Appointments
      summary: Delete an appointment
      description: Deletes an appointment by its ID
      operationId: deleteAppointment
      parameters:
      - name: id
        in: path
        description: Appointment ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Appointment deleted successfully
        "404":
          description: Appointment not found
  /appointments/agenda:
    get:
      tags:
      - Appointments
      summary: View daily agenda
      description: Retrieves appointments for a specific date with optional status
        and service type filters
      operationId: viewDailyAgenda
      parameters:
      - name: date
        in: query
        description: Date to view (ISO format)
        required: true
        schema:
          type: string
          format: date
        example: 2026-02-26
      - name: status
        in: query
        description: Filter by status
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - ACTIVE
          - COMPLETED
          - CANCELED
        example: ACTIVE
      - name: serviceType
        in: query
        description: Filter by service type
        required: false
        schema:
          type: string
          enum:
          - CRECHE
          - HOTEL
        example: CRECHE
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: perPage
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
          format: int32
          default: 20
        example: 20
      - name: sort
        in: query
        description: Sort field
        required: false
        schema:
          type: string
          default: startAt
        example: startAt
      - name: direction
        in: query
        description: Sort direction (asc/desc)
        required: false
        schema:
          type: string
          default: asc
        example: asc
      responses:
        "400":
          description: Invalid date parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationAppointmentResponse"
        "200":
          description: Agenda retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationAppointmentResponse"
components:
  schemas:
...
    CreateAppointmentRequest:
      type: object
      properties:
        pet_id:
          type: string
        service_type:
          type: string
          enum:
          - CRECHE
          - HOTEL
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        notes:
          type: string
    ChangeAppointmentStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - PENDING
          - ACTIVE
          - COMPLETED
          - CANCELED
...
    AppointmentResponse:
      type: object
      properties:
        id:
          type: string
        pet_id:
          type: string
        service_type:
          type: string
          enum:
          - CRECHE
          - HOTEL
        status:
          type: string
          enum:
          - PENDING
          - ACTIVE
          - COMPLETED
          - CANCELED
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaginationAppointmentResponse:
      type: object
      properties:
        currentPage:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentResponse"